Do not try to remove non-existing package information file on rollback
The installation information XML file (components.xml by default) is
not yet written to disk when downloading archives. Canceling the
installation here would try to remove the not-yet-created file
unconditionally.

Task-number: QTIFW-2213

+++ src/libs/installer/packagemanagercore.cpp
@@ -896,11 +896,10 @@ void PackageManagerCore::rollBackInstallation()
             }
 
             d->m_localPackageHub->writeToDisk();
-            if (isInstaller()) {
-                if (d->m_localPackageHub->packageInfoCount() == 0) {
-                    QFile file(d->m_localPackageHub->fileName());
+            if (isInstaller() && d->m_localPackageHub->packageInfoCount() == 0) {
+                QFile file(d->m_localPackageHub->fileName());
+                if (!file.fileName().isEmpty() && file.exists())
                     file.remove();
-                }
             }
 
             if (becameAdmin)
